+++ Automate/IMachine.cs
@@ -4,6 +4,14 @@ namespace Pathoschild.Stardew.Automate
     public interface IMachine : IAutomatable
     {
         /*********
+        ** Accessors
+        *********/
+        /// <summary>A unique ID for the machine type.</summary>
+        /// <remarks>This value should be identical for two machines if they have the exact same behavior and input logic. For example, if one machine in a group can't process input due to missing items, Automate will skip any other empty machines of that type in the same group since it assumes they need the same inputs.</remarks>
+        string MachineTypeID { get; }
+
+
+        /*********
         ** Public methods
         *********/
         /// <summary>Get the machine's processing state.</summary>